+++ lib/kitchen/directions/bake_figure.rb
@@ -4,7 +4,20 @@ module Kitchen
   module Directions
     module BakeFigure
       def self.v1(figure:, number:)
+        return if figure.has_class?('unnumbered') && !figure.has_class?('splash')
+
         figure.wrap(%(<div class="os-figure#{' has-splash' if figure.has_class?('splash')}">))
+        if figure.has_class?('unnumbered') && figure.has_class?('splash')
+          caption = figure.caption&.cut
+          figure.append(sibling:
+            <<~HTML
+              <div class="os-caption-container">
+                #{"<span class=\"os-caption\">#{caption.children}</span>" if caption}
+              </div>
+            HTML
+          )
+          return
+        end
 
         figure.pantry(name: :link_text).store "#{I18n.t(:figure)} #{number}", label: figure.id
         title = figure.title&.cut